fixing up the routes so they give back predictable and useful responses when bad things happen. editing the classes so that it will give empty objects instead of null. editing the launchSettings so that it would startup on the user route.

+++ LightLinkAPI/LightLinkAPI/Controllers/ComputerController.cs
@@ -14,49 +14,59 @@ namespace LightLinkAPI.Controllers
     public class ComputerController : Controller
     {
         public IComputerService ComputerService { get; private set; }
-        public ComputerController(IComputerService computerService)
+        public IUserService UserService { get; private set; }
+
+        public ComputerController(IComputerService computerService, IUserService userService)
         {
             ComputerService = computerService;
+            UserService = userService;
         }
         // GET: api/<controller>
         [HttpGet]
-        public IEnumerable<Computer> Get()
+        public ActionResult Get()
         {
-            return ComputerService.GetAllComputers();
+            return Ok(ComputerService.GetAllComputers());
         }
 
         [HttpGet("unassigned")]
-        public IEnumerable<Computer> GetUnassigned()
+        public ActionResult GetUnassigned()
         {
-            return ComputerService.GetAllUnAssignedComputers();
+            return Ok(ComputerService.GetAllUnAssignedComputers());
         }
 
         // GET api/<controller>/5
         [HttpGet("{username}")]
-        public IEnumerable<Computer> Get(string username)
+        public ActionResult Get(string username)
         {
-            return ComputerService.GetComputersForUser(username);
+            if (!UserService.Exists(username)) return BadRequest("user with the given username does not exist.");
+            return Ok(ComputerService.GetComputersForUser(username));
         }
 
         // POST api/<controller>
         [HttpPost]
-        public void Post([FromBody]Computer value)
+        public ActionResult Post([FromBody]Computer value)
         {
+            if (value is null) return BadRequest(new { error = "not able to create computer object." });
             ComputerService.AddComputer(value);
+            return Ok();
         }
 
         // PUT api/<controller>/5
         [HttpPut("{computername}")]
-        public void Put(string computername, [FromBody]Computer value)
+        public ActionResult Put(string computername, [FromBody]Computer value)
         {
+            if (!ComputerService.Exist(computername)) return BadRequest(new { error = "computer with the given name does not exist." });
             ComputerService.UpdateComputer(computername, value);
+            return Ok();
         }
 
         // DELETE api/<controller>/5
         [HttpDelete("{computername}")]
-        public void Delete(string computername)
+        public ActionResult Delete(string computername)
         {
+            if (!ComputerService.Exist(computername)) return BadRequest(new { error = "computer with given name does not exist." });
             ComputerService.RemoveComputer(computername);
+            return Ok();
         }
     }
 }
